Moving module 'envconfig' to 'file_analysis' to better reflect its purpose
