+++ repronim/file_analysis/__init__.py
@@ -6,7 +6,7 @@
 #   copyright and license terms.
 #
 # ## ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ##
-"""Environment Configuration Class tests
+"""Determine package and distribution information for files
 
 """
