+++ repronim/file_analysis/tests/__init__.py
@@ -6,7 +6,7 @@
 #   copyright and license terms.
 #
 # ## ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ##
-"""ReproNim Experiment Environment Configuration
+"""File analysis tests
 
 """
