+++ repronim/file_analysis/tests/test_configuration.py
@@ -7,10 +7,11 @@
 #
 # ## ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ##
 
-from envconfig.config_item import ConfigurationItem
-from envconfig.configuration import register_config_item 
-from envconfig import configuration
 import pytest
+
+from repronim.file_analysis.config_item import ConfigurationItem
+from repronim.file_analysis import configuration
+from repronim.file_analysis.configuration import register_config_item
 
 # A dict to back up the registered handlers
 backup_handlers = {};
